scraper/requirements.txt
pydantic==2.6.1
requests==2.31.0
selenium==4.17.2
docker==7.0.0
PyYAML==6.0.1
toml==0.10.2
pytest==8.0.0
beautifulsoup4==4.12.3
pytesseract==0.3.10